Add files via upload
In this tableau project as a part of Forage virtual internship I created a tableau dashboard for British Airways reviews where I analyzed the key metrics against aircraft type, countries and by month. I applied filters against the class, continent and travel type. The key takeaway of this project was British Airways needed to improve their in flight services like entertainment and food as it is affecting the overall experience of the passenger and bringing down their overall value for money and hampering their reputation as a premium carrier.