Add files via upload
I created a hotels dashboard for a major aggregator and I gathered the requirements from the stakeholder and created a dashboard on basis of that. I created key metrics for ADR, DSRN, Revpar, Total Revenue, Occupancy %, Realization % etc. The major insight that I gained was hotels business works on Pareto principle and the hotels need to adapt elastic and flexible pricing as it would lead to them generating more revenue during the festive season or during weekends.