Add files via upload
Performed analysis on the sales data of houses in King County, Washington. Designed a calendar widget and used it as a filter against a heatmap,line chart and histogram by comparing the average sales price against metrics like location and month and year in which sale took place. Also used the year in which the house was built,the total buildup area and the carpet area with lawn as filters. Understood how the view is the major deciding factor in deciding the price of the house when compared to the condition of the house.