Add files via upload
This project aims to detect fraudulent credit card transactions and also predict fraudulent transactions. This is highly critical as the fraudulent transactions hamper the reputation of the bank and result in the banks losing the customer base. So, in order to perform the analysis we consider a dataset of credit card transactions over 2 days of Europe credit cardholders.